requirements.txt
# Core
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Memory
redis>=5.0.0
chromadb>=0.5.0
langchain-chroma>=0.1.0

# Voice
openai-whisper>=20231117
speechrecognition>=3.10.0
pyaudio>=0.2.14
pyttsx3>=2.90

# Frontend
streamlit>=1.40.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.5.0
numpy>=1.26.0